Replace data models and test classes
